re place funcs
